07-12-2018

+++ ft_free_tab.c
@@ -6,15 +6,16 @@
 /*   By: aljacque <aljacque@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/07 14:33:57 by aljacque          #+#    #+#             */
-/*   Updated: 2018/12/07 15:11:55 by aljacque         ###   ########.fr       */
+/*   Updated: 2018/12/07 15:23:53 by coremart         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "includes/fillit.h"
+#include <stdlib.h>
 
 void	ft_free_tab(char **tab, size_t size)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < size)
